+++ .gitignore
@@ -40,4 +40,5 @@ yarn-error.log*
 *.tsbuildinfo
 next-env.d.ts
 
+#Backend
 /backend/target/
