+++ src/helper/report-type/report-type-helper.tsx
@@ -1,18 +1,20 @@
 import { ReportType } from "@/components/model/report-type";
 
+export const defaultReportType = {
+    id: '',
+    type: '',
+    label: '',
+    name: '',
+    description: '',
+    primaryTableId: '',
+    primaryTable: '',
+    primaryTableDisplayName: ''
+};
+
 export const useReportTypeInitialValues = (reportType: ReportType | undefined): ReportType => {
     if (reportType) {
         return reportType;
     } else {
-        return {
-            id: '',
-            label: '',
-            name: '',
-            description: '',
-            primaryTableId: '',
-            primaryTable: '',
-            primaryTableDisplayName: ''
-
-        };
+        return defaultReportType;
     }
 }
